[GPU] Consolidate FP16 compression passes for Qwen-Image
Group GPU FP16 compression matchers under a single model pass and
consolidate their implementations and unit tests.

Keep Qwen-Image gated residual and direct Multiply-Sin/Cos patterns
in FP32 to avoid precision loss during FP16 compression.
