+++ src/plugins/intel_gpu/src/plugin/transformations/increase_position_ids_precision.hpp
@@ -56,10 +56,4 @@ public:
     IncreasePositionIdsPrecision();
     bool run_on_model(const std::shared_ptr<ov::Model>& model) override;
 };
-
-class DisableFP16ComForGPTOSSROPEPattern: public ov::pass::MatcherPass {
-public:
-    OPENVINO_MATCHER_PASS_RTTI("DisableFP16ComForROPEPattern");
-    DisableFP16ComForGPTOSSROPEPattern();
-};
 }   // namespace ov::intel_gpu
